ar(feat) finish types split
