+++ lib/types/common.d.ts
@@ -1,0 +1,4 @@
+// common.d.ts
+import type { PostalAddress } from '@types/json-schmea';
+
+export type IAddress = PostalAddress;
